Give the syscall ABI a width, and build ring 3 for a 32-bit machine
arch.bzl marked both 32-bit ports user_flags = None, and the reason was
uabi. Every entry point takes u64 because the ABI is written in 64-bit
words; an argument register here is half that, so a u64 would be passed
in a register pair and shift every argument index the kernel reads --
the shape of the bug that already cost this port a boot handoff (D107).

Refused, not truncated, and not a 181-site sweep either. Retyping every
entry point to usize would have touched 181 call sites across 23
programs whose behaviour only a boot check can verify. The surface stays
u64 and the narrowing happens inside uabi: syscall_arg answers None for a
value that does not fit, and the call returns the kernel-domain
InvalidMapping before it traps. A pointer with its high half silently
removed is a pointer into somebody else's memory, and this tree's rule is
that code which degrades says so -- here that means not degrading at all.

The result widens the other way and is lossless: a failure is six domains
and small codes, and a success on a 32-bit machine is a handle, a count
or an address, all of which that register held already. docs/api/01 says
this now.

Entry points take usize rather than u64, which is not cosmetic: _start
reads its startup argument out of a single register the kernel wrote, and
a u64 parameter would read a pair.

restart-probe builds for riscv32 -- the first compiled ring-3 program for
a 32-bit machine here -- which gives D258's Machine::RiscV32 a producer.
//tools/checks:user_elf32_test feeds the artifact a real linker produced
to the real kcore::elf parser: the entry is where user-riscv32.ld links,
the text segment is read+execute and not writable, and the same bytes are
refused for the 64-bit target, which only the class byte can do since
RISC-V gives both widths one machine number. A gate rather than a unit
test, because it is the join between two packages that never reference
each other. The inversion fires against the real artifact: reading ELF32
headers at the ELF64 offsets refuses the toolchain's own output.

Nothing runs the program yet. kernel-riscv32 reaches ring 3 through a
bespoke two-syscall trap handler with no Executive, no process table and
no kcore::dispatch. arm32 keeps user_flags = None, because the same
decision needs its own kernel side to be worth making. And uabi::layout
is still per-port cfgs naming no 32-bit machine.

Ledger row D259.

+++ userspace/restart-probe/src/main.rs
@@ -29,12 +29,18 @@ const SYS_PROCESS_EXIT: u64 = 5;
 
 /// The ELF entry point. `arg` is what the supervisor passed in
 /// `ProcessStartArgs::arg`: the code to exit with.
+///
+/// **`usize` rather than `u64`**, which matters on exactly one kind of machine
+/// and matters completely there: the kernel hands this over in a single
+/// argument register, and on a 32-bit port a `u64` parameter is passed in a
+/// register *pair* — so the program would read its startup argument out of two
+/// registers the kernel wrote one of (build/README.md, D259).
 // SAFETY: `no_mangle` gives this function the name the linker script's ENTRY
 // resolves, which is what makes it the ELF's entry point. Nothing else in the
 // program is exported, so there is no symbol to collide with.
 #[unsafe(no_mangle)]
-pub extern "C" fn _start(arg: u64) -> ! {
-    syscall2(SYS_PROCESS_EXIT, arg, 0);
+pub extern "C" fn _start(arg: usize) -> ! {
+    syscall2(SYS_PROCESS_EXIT, arg as u64, 0);
     // The kernel does not return from an exit; spin rather than fall off the
     // end of the entry point if a future one ever did.
     loop {
